pyproject.toml
[project]
name = "hey-to-gmail"
version = "1.0.0"
description = "Import emails from HEY MBOX export to Gmail with label management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.20.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.scripts]
hey-to-gmail = "hey_to_gmail.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
